#501: dir items
