+++ lib/zold/commands/node.rb
@@ -385,7 +385,7 @@ module Zold
       end
 
       def info(msg)
-        @log.debug('WEBRICK ' + msg)
+        @log.debug("WEBRICK #{msg}")
       end
 
       def debug(msg)
@@ -393,11 +393,11 @@ module Zold
       end
 
       def error(msg)
-        @log.error('WEBRICK ' + msg)
+        @log.error("WEBRICK #{msg}")
       end
 
       def fatal(msg)
-        @log.error('WEBRICK ' + msg)
+        @log.error("WEBRICK #{msg}")
       end
 
       def debug?
